Fetch basic Facebook info and add logout

+++ public/js/facebook.js
@@ -5,18 +5,14 @@ function checkLoginState() {
 }
 
 function statusChangeCallback(response) {
-  console.log('Facebook login status changed.');
   if (response.status === 'connected') {
-    console.log('Successfully logged in with Facebook');
     // const userId = response.authResponse.userID;
     // const accessToken = response.authResponse.accessToken;
-    fetchUserInfo();
-    // $.get(`https://graph.facebook.com${userId}?access_token=${accessToken}&fields=email,name,picture`);
-    // FB.api('/me?fields=name,first_name,picture.width(480)');
+    setUserInfo();
   }
 }
 
-function fetchUserInfo() {
+function setUserInfo() {
   FB.api('/me?fields=name,first_name,email,picture.width(100)', function (response) {
     if (response && !response.error) {
       window.localStorage.setItem('id', response.id);
@@ -24,13 +20,29 @@ function fetchUserInfo() {
       window.localStorage.setItem('first_name', response.first_name);
       window.localStorage.setItem('email', response.email);
       window.localStorage.setItem('picture', response.picture.data.url);
+      window.location.href = "/home";
     }
   });
 }
 
+// function fetchUserInfo() {
+//   const userObj = new Object();
+//   const localStorage = window.localStorage;
+//   userObj.id = localStorage.getItem('id');
+//   userObj.name = localStorage.getItem('name');
+//   userObj.first_name = localStorage.getItem('first_name');
+//   userObj.email = localStorage.getItem('email');
+//   userObj.picture = localStorage.getItem('picture');
+//   return userObj;
+// }
+
 function logoutFacebook() {
-  FB.logout(function (response) {
-    console.log('Logged out of Facebook', response);
-    window.localStorage.clear();
-  })
+  FB.getLoginStatus(function (response) {
+    if (response.status === 'connected') {
+      FB.logout(function (res) {
+        window.localStorage.clear();
+        window.location.href = "/login";
+      });
+    }
+  });
 }
